test: Use nix run github:flox/flox
Test against HEAD of the `main` branch so that we can exercise new
fixes, features, and catch regressions. This is needed right now to
unblock the PRs for pure manifest builds which depend on a fix that has
just been merged in `flox/flox`.

It doesn't quite feel right defining the default in two places but I'm
slightly uncomfortable with only declaring it in one and it matches the
existing style.